+++ src/controllers/SlideoutController.php
@@ -19,23 +19,21 @@ class SlideoutController extends Controller
     public function actionRender(): Response
     {
         $customer = $this->request->getRequiredParam('customer');
+        $linkedCustomers = PluginSales::$plugin->settings->getLinkedCustomers($customer);
         $url = $this->getUrlFromCustomer($customer);
         $sales = PluginSales::$plugin->reports->getSalesData($customer);
 
         return $this->asCpScreen()
             ->contentTemplate('plugin-sales/_slideout', [
                 'customer' => $customer,
+                'linkedCustomers' => $linkedCustomers,
                 'url' => $url,
                 'sales' => $sales,
             ]);
     }
 
-    private function getUrlFromCustomer(?string $customer): ?string
+    private function getUrlFromCustomer(string $customer): ?string
     {
-        if ($customer === null) {
-            return null;
-        }
-
         $domain = explode('@', $customer)[1] ?? null;
 
         if ($domain === null) {