+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes for Plugin Sales
 
+## 3.1.0 - 2024-04-26
+
+### Added
+
+- Added the ability to link customers in the plugin settings.
+
+### Changed
+
+- Saving the plugin settings no longer automatically refreshes all plugin sales. Instead, a link is provided to perform a full refresh.
+
 ## 3.0.1 - 2024-04-23
 
 ### Fixed